Developed a predictive model aimed at identifying fraudulent transactions within the operations of a financial entity.
The objective of this endeavor is to create a machine-learning algorithm capable of forecasting fraudulent transactions within the operations of a financial institution. The dataset for this endeavor is provided in CSV format, comprising 6,362,620 entries distributed across 10 columns. The model will undergo estimation using calibration data and subsequent validation testing. This undertaking necessitates a combination of statistical analysis and creative judgment.